pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "airline-sked"
version = "0.1.0"
description = "한-일 항공 노선 스케줄 변경 감지 및 알림 시스템"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "sqlmodel>=0.0.16",
    "aiosqlite>=0.20",
    "httpx>=0.27",
    "selectolax>=0.3",
    "beautifulsoup4>=4.12",
    "playwright>=1.42",
    "apscheduler>=3.10",
    "python-telegram-bot>=21",
    "discord.py>=2.3",
    "pydantic-settings>=2.1",
    "typer[all]>=0.9",
    "rich>=13",
    "tenacity>=8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "ruff>=0.3",
    "mypy>=1.8",
    "pre-commit>=3.6",
]

[project.scripts]
airline-sked = "airline_sked.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/airline_sked"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
